docs: remove `@Nls` from all `Message` types
`@Nls` should be added in `Component.text(String)` methods, it is not orefield-text's responsibility to do this by annotating `Message`, which also goes against the convention of using `@Nls` on `java.lang.String`s.

The legacy string version may annotate the `Message` type parameter with `@Nls` to signify that all strings are user-visible.